Go back to original package structure
and just nest the cli
