+++ versioneer/base_name_test.go
@@ -1,7 +1,7 @@
 package versioneer_test
 
 import (
-	"github.com/kairos-io/kairos-sdk/versioneer/pkg/versioneer"
+	"github.com/kairos-io/kairos-sdk/versioneer"
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 )
