+++ versioneer/cli.go
