+++ versioneer/container_name_test.go
@@ -1,14 +1,15 @@
 package versioneer_test
 
 import (
-	"github.com/kairos-io/kairos-sdk/versioneer/pkg/versioneer"
+	"github.com/kairos-io/kairos-sdk/versioneer"
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 )
 
-var _ = Describe("BootableName", func() {
+var _ = Describe("ContainerName", func() {
 	var artifact versioneer.Artifact
 	var expectedName string
+	var registryAndOrg string
 
 	BeforeEach(func() {
 		artifact = versioneer.Artifact{
@@ -20,16 +21,19 @@ var _ = Describe("BootableName", func() {
 			Version:         "v2.4.2",
 			SoftwareVersion: "k3sv1.26.9+k3s1",
 		}
+
+		registryAndOrg = "quay.io/kairos"
 	})
 
 	When("artifact is valid", func() {
 		When("SoftwareVersion is empty", func() {
 			BeforeEach(func() {
 				artifact.SoftwareVersion = ""
-				expectedName = "kairos-opensuse-leap-15.5-standard-amd64-generic-v2.4.2"
+				expectedName = "quay.io/kairos/opensuse:leap-15.5-standard-amd64-generic-v2.4.2"
 			})
+
 			It("returns the name", func() {
-				name, err := artifact.BootableName()
+				name, err := artifact.ContainerName(registryAndOrg)
 				Expect(err).ToNot(HaveOccurred())
 				Expect(name).To(Equal(expectedName))
 			})
@@ -37,10 +41,10 @@ var _ = Describe("BootableName", func() {
 
 		When("SoftwareVersion is not empty", func() {
 			BeforeEach(func() {
-				expectedName = "kairos-opensuse-leap-15.5-standard-amd64-generic-v2.4.2-k3sv1.26.9+k3s1"
+				expectedName = "quay.io/kairos/opensuse:leap-15.5-standard-amd64-generic-v2.4.2-k3sv1.26.9-k3s1"
 			})
 			It("returns the name", func() {
-				name, err := artifact.BootableName()
+				name, err := artifact.ContainerName(registryAndOrg)
 				Expect(err).ToNot(HaveOccurred())
 				Expect(name).To(Equal(expectedName))
 			})
@@ -49,11 +53,11 @@ var _ = Describe("BootableName", func() {
 
 	When("artifact is invalid", func() {
 		BeforeEach(func() {
-			artifact.Version = ""
+			artifact.Flavor = ""
 		})
 		It("returns an error", func() {
-			_, err := artifact.BootableName()
-			Expect(err).To(MatchError("Version is empty"))
+			_, err := artifact.ContainerName(registryAndOrg)
+			Expect(err).To(MatchError("Flavor is empty"))
 		})
 	})
 })
