+++ versioneer/registry_inspector.go
